CleanUp DetailsController, Create DetailsService, ...
